% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date.utils.r
\name{apply.business.days}
\alias{apply.business.days}
\title{Apply Business Date logic to provided Dates function}
\usage{
apply.business.days(dates, dates.fn = NULL, calendar = NULL, base = T)
}
\arguments{
\item{dates}{dates, dates are assumed to contain only business dates}

\item{dates.fn}{dates function that returns index of selected dates, \strong{defaults to NULL}}

\item{calendar}{RQuantLib calendar name to use to determine business days, \strong{defaults to NULL}}
}
\value{
index of business days
}
\description{
Apply Business Date logic to provided Dates function
}
\examples{
\dontrun{
dates = seq(Sys.Date()-1000, Sys.Date(), 1)
apply.business.days(dates, date.week.ends, 'UnitedStates/NYSE')
apply.business.days(dates, date.week.ends, 'Canada/TSX')
}
}

