% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date.utils.r
\name{business.days}
\alias{business.days}
\title{Extract Business Days}
\usage{
business.days(from = Sys.Date(), to = as.Date(from) + 31, holidays = NULL,
  dates = NULL, calendar = NULL, return.index = F)
}
\arguments{
\item{from}{start date (ignored if dates are provided)}

\item{to}{start date (ignored / optional if dates are provided)}

\item{holidays}{list of holidays, \strong{defaults to NULL}}

\item{dates}{dates, \strong{defaults to NULL}}

\item{calendar}{RQuantLib calendar name to use to determine business days, \strong{defaults to NULL}}

\item{return.index}{flag to return index of valid dates instead of dates, \strong{defaults to FALSE}}
}
\value{
business days
}
\description{
Extract Business Days
}
\examples{
\dontrun{
library(RQuantLib)
from = as.Date('10Jun2013','\%d\%b\%Y')
to = as.Date('10Jan2014','\%d\%b\%Y')
holidays = getHolidayList('UnitedStates/NYSE', from, to)
holidays = getHolidayList('Canada/TSX', from, to)
business.days(from, to, holidays = holidays)

business.days(dates = dates, calendar = 'UnitedStates/NYSE')
}
}

