% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date.utils.r
\name{business.days.till.end}
\alias{business.days.since.end}
\alias{business.days.till.end}
\alias{last.business.day}
\alias{next.business.day}
\title{Business Date Functions}
\usage{
business.days.till.end(from, holidays = NULL, fn.ends = date.month.ends)

business.days.since.end(from, holidays = NULL, fn.ends = date.month.ends)

next.business.day(from, holidays = NULL, offset = 0)

last.business.day(from, holidays = NULL, offset = 0)
}
\arguments{
\item{from}{start date}

\item{holidays}{list of holidays, \strong{defaults to NULL}}

\item{fn.ends}{function that return periods ends, \strong{defaults to date.month.ends}}
}
\value{
number of business days
}
\description{
Business Date Functions
}
\examples{
\dontrun{
from = as.Date('27Dec2013','\%d\%b\%Y')
holidays = getHolidayList('UnitedStates/NYSE', from-40, from+10)
business.days.till.end(from, holidays)
business.days.since.end(from, holidays)
}
}

