% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/custom.date.r
\name{custom.date.bus}
\alias{custom.date.bus}
\title{Custom Date function with Business Date logic}
\usage{
custom.date.bus(expr, dates, calendar = NULL)
}
\arguments{
\item{expr}{expression that describes the dates to be extracted}

\item{dates}{dates}

\item{calendar}{RQuantLib calendar name to use to determine business days, \strong{defaults to NULL}}
}
\value{
list with days.since and days.till arrays
}
\description{
Custom Date function with Business Date logic
}
\examples{
\dontrun{
dates = seq(Sys.Date()-1000, Sys.Date(), 1)
custom.date.bus('last day in Apr', dates, 'UnitedStates/NYSE')
custom.date.bus('last day in Apr', dates, 'Canada/TSX')
}
}

