% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/custom.date.r
\name{custom.date}
\alias{custom.date}
\title{Custom Date function}
\usage{
custom.date(expr, dates)
}
\arguments{
\item{expr}{expression that describes the dates to be extracted}

\item{dates}{dates}
}
\value{
list with days.since and days.till arrays
}
\description{
Custom Date function
}
\examples{
\dontrun{
dates = seq(Sys.Date()-1000, Sys.Date(), 1)
custom.date('last day in Apr', dates)
}
}

