% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date.utils.r
\name{date.ends.index}
\alias{date.ends.index}
\title{out is result of the business.days.location.end}
\usage{
date.ends.index(out, timing)
}
\description{
out is result of the business.days.location.end
}

