% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date.utils.r
\name{date.ends}
\alias{date.ends}
\title{Date Ends function}
\usage{
date.ends(dates, periodicity, by = 1, skip = 0, last.date = T,
  calendar = NULL)
}
\arguments{
\item{dates}{dates}

\item{periodicity}{periodicity}

\item{by}{take every *by* item, \strong{defaults to 1}}

\item{skip}{skip periods, \strong{defaults to 0}}

\item{last.date}{flag to include last date, \strong{defaults to TRUE}}

\item{calendar}{RQuantLib calendar name to use to determine business days, \strong{defaults to NULL}}
}
\value{
index of date ends
}
\description{
Date Ends function
}
\examples{
\dontrun{
dates = seq(Sys.Date()-1000, Sys.Date(), 1)
date.ends(dates, 'year', calendar = 'UnitedStates/NYSE')
date.ends(dates, 'year', calendar = 'Canada/TSX')
date.ends(dates, 'bi-weekly')
}
}

