% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date.utils.r
\name{date.dayofweek}
\alias{date.all}
\alias{date.day}
\alias{date.dayofweek}
\alias{date.month}
\alias{date.quarter}
\alias{date.semiannual}
\alias{date.week}
\alias{date.year}
\title{Dates Functions}
\usage{
date.dayofweek(dates)

date.day(dates)

date.week(dates)

date.month(dates)

date.quarter(dates)

date.semiannual(dates)

date.year(dates)

date.all(dates)
}
\arguments{
\item{dates}{collection of dates}
}
\value{
transformed dates
}
\description{
Dates Functions
}
\examples{
\dontrun{
date.dayofweek(Sys.Date())
}
}

