% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date.utils.r
\name{date.week.ends}
\alias{date.end}
\alias{date.month.ends}
\alias{date.quarter.ends}
\alias{date.semiannual.ends}
\alias{date.week.ends}
\alias{date.year.ends}
\title{Dates Index Functions}
\usage{
date.week.ends(dates, last.date = T)

date.month.ends(dates, last.date = T)

date.quarter.ends(dates, last.date = T)

date.semiannual.ends(dates, last.date = T)

date.year.ends(dates, last.date = T)

date.end(date = Sys.Date(), periodicity = "months",
  date.format = "\%Y-\%m-\%d")
}
\arguments{
\item{dates}{collection of dates}

\item{last.date}{flag to include last date, \strong{defaults to TRUE}}
}
\value{
location of the week/month/year ends
}
\description{
Dates Index Functions

last calendar day of period
date.end('2014-01-13')
}
\examples{
\dontrun{
date.week.ends(seq(Sys.Date()-100, Sys.Date(), 1))
}
}

