% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/date.utils.r
\name{date.periodicity.map}
\alias{date.ends.fn}
\alias{date.periodicity.map}
\title{Dates Mapping Functions}
\usage{
date.periodicity.map(periodicity)

date.ends.fn(periodicity)
}
\arguments{
\item{periodicity}{periodicity (i.e. weeks, months)}
}
\value{
standard periodicity
}
\description{
Dates Mapping Functions
}
\examples{
\dontrun{
date.freq.map('week')
}
}

